+++ astroquery/heasarc/tests/test_heasarc_remote.py
@@ -10,6 +10,34 @@ from ...utils import commons
 
 @pytest.mark.remote_data
 class TestHeasarc:
+    def test_custom_args(self):
+        object_name = 'Crab'
+        mission = 'intscw'
+
+        heasarc = Heasarc()
+
+        table = heasarc.query_object(object_name,
+                                     mission=mission,
+                                     radius='1 degree',
+                                     time="2020-09-01 .. 2020-12-01",
+                                     resultmax=10,
+                                     good_isgri=">1000",
+                                     )
+
+    def test_filter_custom_args(self):
+        object_name = 'Crab'
+        mission = 'intscw'
+
+        heasarc = Heasarc()
+
+        with pytest.raises(ValueError):
+            table = heasarc.query_object(object_name,
+                                         mission=mission,
+                                         radius='1 degree',
+                                         time="2020-09-01 .. 2020-12-01",
+                                         resultmax=10,
+                                         very_good_isgri=">1000",
+                                         )
 
     def test_basic_example(self):
         mission = 'rosmaster'