+++ CHANGES.rst
@@ -8,6 +8,17 @@ New Tools and Services
 Service fixes and enhancements
 ------------------------------
 
+heasarc
+^^^^^^^
+
+- Added posibility to query limited time range. [#1988]
+- Add alternative instance of HEASARC Server, maintained by 
+  INTEGRAL Science Data Center. [#1988]
+
+
+Service fixes and enhancements
+------------------------------
+
 
 
 0.4.2 (2021-05-14)